requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto==2.49.0
boto3==1.26.151
botocore==1.29.151
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
linode-cli==5.40.0
markdown-it-py==2.2.0
mdurl==0.1.2
packaging==23.1
Pillow==9.5.0
Pygments==2.15.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.4.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
